Rename folder movies >> dashboard
