+++ state/dashboard/reducer.js
@@ -1,0 +1,37 @@
+import * as types from './types';
+
+const initialState = {
+  loading: false,
+  error: '',
+  ids: [],
+  totalItems: 0,
+  currentPage: 0
+};
+
+const dashboard = (state = initialState, action) => {
+  switch (action.type) {
+    case types.FETCH_TRENDING_MOVIES:
+      return {
+        ...state,
+        loading: true
+      };
+    case types.FETCH_TRENDING_MOVIES_SUCCESS:
+      return {
+        ...state,
+        loading: false,
+        ids: action.ids,
+        totalItems: action.totalItems,
+        currentPage: action.currentPage
+      };
+    case types.FETCH_TRENDING_MOVIES_ERROR:
+      return {
+        ...state,
+        loading: false,
+        error: action.error
+      };
+    default:
+      return state;
+  }
+};
+
+export default dashboard;
