+++ state/dashboard/types.js
@@ -1,0 +1,3 @@
+export const FETCH_TRENDING_MOVIES = 'FETCH_TRENDING_MOVIES';
+export const FETCH_TRENDING_MOVIES_SUCCESS = 'FETCH_TRENDING_MOVIES_SUCCESS';
+export const FETCH_TRENDING_MOVIES_ERROR = 'FETCH_TRENDING_MOVIES_ERROR';
