--- state/movies/types.js
@@ -1,3 +1,0 @@
-export const FETCH_TRANDING_MOVIES = 'FETCH_TRANDING_MOVIES';
-export const FETCH_TRANDING_MOVIES_SUCCESS = 'FETCH_TRANDING_MOVIES_SUCCESS';
-export const FETCH_TRANDING_MOVIES_ERROR = 'FETCH_TRANDING_MOVIES_ERROR';
