+++ state/store/reducer.js
@@ -1,12 +1,12 @@
 import { combineReducers } from 'redux';
 import data from '../data/reducer';
 import session from '../session/reducer';
-import trendingMovies from '../movies/reducer';
+import dashboard from '../dashboard/reducer';
 
 const reducer = combineReducers({
   data,
   session,
-  trendingMovies
+  dashboard
 });
 
 export default reducer;
